Added qtproject scaffold
The qtproject generates a simple Qt project, backed by the Autotools.
The project uses the qtmain scaffold to build a "Hello, world!"-style
application, as well as setting up the scaffold for some tests.

The qtproject will be under version control, with everything already
committed when the scaffold is complete. .gitignore files have also been
set up so that any Autotools files are ignored.